agents/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-builder-agents"
version = "0.1.0"
description = "Phase 0 ARQ worker baseline for the AI Resume Builder"
requires-python = ">=3.9"
dependencies = [
  "arq>=0.26.3",
  "httpx>=0.28.1",
  "langchain-openai>=0.3.35",
  "playwright>=1.58.0",
  "pydantic-settings>=2.8.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0"
]

[tool.setuptools]
py-modules = ["worker", "generation", "validation", "assembly"]
